refactor(ale_claw): Tier 4 — extract agent_loop_helpers.py (stateless free fns)
Moves the module-level free functions out of agent_loop.py into
agent_loop_helpers.py: _maybe_sanitize_screenshot, _rewrite_input_image_to_image_url,
has_done_signal (+ _DONE_LINE_RE), _strip_images_from_messages (+ _IMAGE_BLOCK_TYPES),
_extract_messages_for_compaction. These touch no instance state; the helper module
depends on _message_shapes + image_sanitization (and SessionManager only as a
TYPE_CHECKING annotation). agent_loop imports the ones the class uses back, which
also re-exports has_done_signal (deployer.py imports it) and _strip_images_from_messages
(tests import it) — all `from .agent_loop import ...` paths unchanged. Dropped
agent_loop's now-unused base64/re/image_sanitization imports. The orchestrator class
OpenClawComputerAgent is intentionally left whole (cohesive stateful lifecycle; a
mixin teardown would fragment it for little readability gain). agent_loop.py 1267 ->
1115. Pure move; full suite green (928 passed, 3 skipped).
